// lib/api-zod/src/generated/types/staffRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Blessing Funeral Scheme operations API
 * OpenAPI spec version: 0.1.0
 */

export type StaffRole = typeof StaffRole[keyof typeof StaffRole];


export const StaffRole = {
  administrator: 'administrator',
  manager: 'manager',
  coordinator: 'coordinator',
  support: 'support',
} as const;